+++ sdk/metric/meter_test.go
@@ -31,7 +31,6 @@ import (
 	"go.opentelemetry.io/otel/internal/global"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
-	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata/metricdatatest"
 	"go.opentelemetry.io/otel/sdk/resource"
@@ -743,6 +742,11 @@ func TestMeterCreatesInstrumentsValidations(t *testing.T) {
 }
 
 func TestValidateInstrumentName(t *testing.T) {
+	const longName = "longNameOver255characters" +
+		"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx" +
+		"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx" +
+		"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx" +
+		"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
 	testCases := []struct {
 		name string
 
@@ -776,8 +780,8 @@ func TestValidateInstrumentName(t *testing.T) {
 			wantErr: fmt.Errorf("%w: name!: must only contain [A-Za-z0-9_.-]", ErrInstrumentName),
 		},
 		{
-			name:    "someverylongnamewhichisover63charactersbutallofwhichmatchtheregexp",
-			wantErr: fmt.Errorf("%w: someverylongnamewhichisover63charactersbutallofwhichmatchtheregexp: longer than 63 characters", ErrInstrumentName),
+			name:    longName,
+			wantErr: fmt.Errorf("%w: %s: longer than 255 characters", ErrInstrumentName, longName),
 		},
 	}
 
@@ -1132,8 +1136,8 @@ func TestUnregisterUnregisters(t *testing.T) {
 }
 
 func TestRegisterCallbackDropAggregations(t *testing.T) {
-	aggFn := func(InstrumentKind) aggregation.Aggregation {
-		return aggregation.Drop{}
+	aggFn := func(InstrumentKind) Aggregation {
+		return AggregationDrop{}
 	}
 	r := NewManualReader(WithAggregationSelector(aggFn))
 	mp := NewMeterProvider(WithReader(r))
@@ -1813,11 +1817,11 @@ func BenchmarkInstrumentCreation(b *testing.B) {
 	}
 }
 
-func testNilAggregationSelector(InstrumentKind) aggregation.Aggregation {
+func testNilAggregationSelector(InstrumentKind) Aggregation {
 	return nil
 }
-func testDefaultAggregationSelector(InstrumentKind) aggregation.Aggregation {
-	return aggregation.Default{}
+func testDefaultAggregationSelector(InstrumentKind) Aggregation {
+	return AggregationDefault{}
 }
 func testUndefinedTemporalitySelector(InstrumentKind) metricdata.Temporality {
 	return metricdata.Temporality(0)